Draw separately lines in different directions of the same road

+++ src/main/scala/com/simulator/common/Position.scala
@@ -6,8 +6,8 @@ object Position {
   def distance(p1: Position, p2: Position) =
     sqrt(pow(p1.x - p2.x, 2) + pow(p1.y - p2.y, 2))
 
-  def interpolate(p1: Position, p2: Position)(alpha: Float): Position =
+  def interpolate(p1: Position, p2: Position)(alpha: Double): Position =
     Position(p1.x * alpha + p2.x * (1 - alpha), p1.y * alpha + p2.y * (1 - alpha))
 }
 
-case class Position(x: Float, y: Float)
+case class Position(x: Double, y: Double)
