+++ src/main/scala/com/simulator/visualization/Rect.scala
@@ -1,8 +1,8 @@
 package com.simulator.visualization
 
 object Rect {
-  def fromLTRB(left: Float, top: Float, right: Float, bottom: Float) =
+  def fromLTRB(left: Double, top: Double, right: Double, bottom: Double) =
     Rect(left, top, right - left, bottom - top)
 }
 
-case class Rect(x: Float, y: Float, w: Float, h: Float)
+case class Rect(x: Double, y: Double, w: Double, h: Double)
